AI-Medical-Image-Analysis-System/requirements.txt
streamlit>=1.28.0
tensorflow==2.18.0
numpy>=1.26.0
Pillow==10.0.1
reportlab==4.0.4
plotly==5.17.0
pandas>=2.0.3
python-dotenv==1.0.0
google-generativeai==0.8.3